внесены изменения после проверки линтеров
