+++ markup/add-new-point.html
@@ -103,7 +103,7 @@
                     </label>
                     <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
 
-                    <div class="event__type-list">
+                    <div class="event__type-list"> <!-- этот блок в отдельный компонент event-type-list !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
                       <fieldset class="event__type-group">
                         <legend class="visually-hidden">Event type</legend>
